bom creation

+++ database/migrations/2024_11_08_091013_create_prd_bill_of_material_childs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->integer('parent_id');
             $table->string('item_code');
             $table->string('item_description');
-            $table->integer('quantity');
+            $table->double('quantity');
             $table->string('measure');
             $table->timestamps();
         });
